requirements.txt
pdfplumber
torch
transformers
sentencepiece
accelerate
llama-cpp-python